[Refactor] Chat -> Channel, Message 로 분리
