+++ src/types/message.type.ts
@@ -1,0 +1,14 @@
+import { User } from '@/types/user.type';
+
+type MessageTypes = 'image' | 'text';
+
+export interface SendMessage {
+  type: MessageTypes;
+  content: string;
+  user: User;
+  channelId: string;
+}
+
+export interface ReceiveMeesage extends SendMessage {
+  date: string;
+}
